JS-3003: date methods -> UtilDate
