+++ src/ts/component/menu/item/filter.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { I, Relation, UtilCommon, translate } from 'Lib';
+import { I, Relation, UtilCommon, translate, UtilDate } from 'Lib';
 import { Icon, Tag, IconObject } from 'Component';
 import { detailStore } from 'Store';
 import { SortableHandle, SortableElement } from 'react-sortable-hoc';
@@ -55,7 +55,7 @@ const MenuItemFilter = observer(class MenuItemFilter extends React.Component<Pro
 				let name = String(filterOption.name || '').toLowerCase();
 
 				if (quickOption == I.FilterQuickOption.ExactDate) {
-					v.push(value !== null ? UtilCommon.date('d.m.Y', value) : '');
+					v.push(value !== null ? UtilDate.date('d.m.Y', value) : '');
 				} else
 				if ([ I.FilterQuickOption.NumberOfDaysAgo, I.FilterQuickOption.NumberOfDaysNow ].includes(quickOption)) {
 					value = Number(value) || 0;
