+++ src/ts/component/page/auth/deleted.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Frame, Title, Label, Error, Header, Button } from 'Component';
-import { I, UtilCommon, UtilRouter, C, Action, Survey, UtilObject, analytics, translate } from 'Lib';
+import { I, UtilCommon, UtilRouter, C, Action, Survey, UtilObject, analytics, translate, UtilDate } from 'Lib';
 import { authStore, popupStore } from 'Store';
 import { observer } from 'mobx-react';
 import { PieChart } from 'react-minimal-pie-chart';
@@ -36,7 +36,7 @@ const PageAuthDeleted = observer(class PageAuthDeleted extends React.Component<I
 		};
 
 		const { error } = this.state;
-		const duration = Math.max(0, account.status.date - UtilCommon.time());
+		const duration = Math.max(0, account.status.date - UtilDate.now());
 		const days = Math.max(1, Math.floor(duration / 86400));
 		const dt = `${days} ${UtilCommon.plural(days, translate('pluralDay'))}`;
