+++ src/ts/component/popup/page/settings/space/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Icon, Title, Label, Input, IconObject, Button, ProgressBar } from 'Component';
-import { UtilObject, UtilMenu, UtilCommon, UtilData, UtilFile, I, translate, Renderer, Preview, analytics } from 'Lib';
+import { UtilObject, UtilMenu, UtilCommon, UtilData, UtilFile, I, translate, Renderer, Preview, analytics, UtilDate } from 'Lib';
 import { observer } from 'mobx-react';
 import { detailStore, menuStore, commonStore, authStore, blockStore } from 'Store';
 import Constant from 'json/constant.json';
@@ -48,7 +48,7 @@ const PopupSettingsSpaceIndex = observer(class PopupSettingsSpaceIndex extends R
 					<div className="sides">
 						<div className="side left">
 							<Title text={translate(`popupSettingsSpaceIndexCreationDateTitle`)} />
-							<Label text={UtilCommon.date(UtilData.dateFormat(I.DateFormat.Short), space.createdDate)} />
+							<Label text={UtilDate.date(UtilData.dateFormat(I.DateFormat.Short), space.createdDate)} />
 						</div>
 					</div>
 				</div>
