+++ src/ts/lib/survey.ts
@@ -1,4 +1,4 @@
-import { I, Storage, UtilCommon, analytics, Renderer, translate, UtilObject, UtilData } from 'Lib';
+import { I, Storage, UtilCommon, analytics, Renderer, translate, UtilObject, UtilData, UtilDate } from 'Lib';
 import { popupStore, authStore } from 'Store';
 import Surveys from 'json/survey.json';
 
@@ -42,7 +42,7 @@ class Survey {
 				break;
 
 			case I.SurveyType.Pmf:
-				param.time = UtilCommon.time();
+				param.time = UtilDate.now();
 				break;
 		};
 
@@ -61,7 +61,7 @@ class Survey {
 
 			case I.SurveyType.Pmf:
 				param.cancel = true;
-				param.time = UtilCommon.time();
+				param.time = UtilDate.now();
 				break;
 		};
 
@@ -74,7 +74,7 @@ class Survey {
 	};
 
 	checkPmf () {
-		const time = UtilCommon.time();
+		const time = UtilDate.now();
 		const obj = Storage.getSurvey(I.SurveyType.Pmf);
 		const timeRegister = Number(Storage.get('timeRegister')) || 0;
 		const lastCompleted = Number(obj.time || Storage.get('lastSurveyTime')) || 0;
@@ -102,7 +102,7 @@ class Survey {
 	checkRegister () {
 		const timeRegister = Number(Storage.get('timeRegister')) || 0;
 		const isComplete = this.isComplete(I.SurveyType.Register);
-		const surveyTime = timeRegister && ((UtilCommon.time() - 86400 * 7 - timeRegister) > 0);
+		const surveyTime = timeRegister && ((UtilDate.now() - 86400 * 7 - timeRegister) > 0);
 
 		if (!isComplete && surveyTime && !popupStore.isOpen()) {
 			this.show(I.SurveyType.Register);
